Consolidated code models
