+++ trueseeing/core/flow/data.py
@@ -27,7 +27,7 @@ from trueseeing.core.ui import ui
 if TYPE_CHECKING:
   from typing import List, Callable, TypeVar, Type, Any, Iterable, Mapping, Set, Optional, FrozenSet, Union
   from trueseeing.core.store import Store
-  from trueseeing.core.code.op import Op, Token
+  from trueseeing.core.code.model import Op, Token
 
 class DataFlows:
   class NoSuchValueError(Exception):
