+++ trueseeing/core/literalquery.py
@@ -22,7 +22,7 @@ import os.path
 import pkg_resources
 
 import trueseeing.core
-from trueseeing.core.code.op import Op
+from trueseeing.core.code.model import Op
 
 if TYPE_CHECKING:
   from typing import Any, Iterable, Tuple, Dict, Optional
