+++ trueseeing/core/store.py
@@ -27,7 +27,7 @@ from trueseeing.core.literalquery import Query
 
 if TYPE_CHECKING:
   from typing import Optional, List, Any
-  from trueseeing.core.code.op import Op, Token
+  from trueseeing.core.code.model import Op, Token
 
 class Store:
   def __init__(self, path: str) -> None:
